+++ pkg/webui/console/views/application-collaborators/index.js
@@ -54,7 +54,7 @@ export default class ApplicationCollaborators extends React.Component {
     const { match } = this.props
 
     return (
-      <ErrorView ErrorComponent={SubViewError}>
+      <ErrorView errorRender={SubViewError}>
         <Switch>
           <Route exact path={`${match.path}`} component={ApplicationCollaboratorsList} />
           <Route exact path={`${match.path}/add`} component={ApplicationCollaboratorAdd} />