+++ pkg/webui/console/views/gateway-api-keys/index.js
@@ -51,7 +51,7 @@ export default class GatewayApiKeys extends React.Component {
     const { match } = this.props
 
     return (
-      <ErrorView ErrorComponent={SubViewError}>
+      <ErrorView errorRender={SubViewError}>
         <Switch>
           <Route exact path={`${match.path}`} component={GatewayApiKeysList} />
           <Route exact path={`${match.path}/add`} component={GatewayApiKeyAdd} />